Update to support long version codes.
Includes new minimal tests for installing apps with major version codes.

Bug: 64459786
Test: cts-tradefed run cts-dev --primary-abi-only --skip-device-info --skip-preconditions --skip-connectivity-check --module CtsAppSecurityHostTestCases --test android.appsecurity.cts.MajorVersionTest

+++ hostsidetests/os/test-apps/StaticSharedLibConsumerApp1/src/android/os/lib/consumer1/UseSharedLibraryTest.java
@@ -187,7 +187,7 @@ public class UseSharedLibraryTest {
         // Make sure we see the lib we depend on via getting installed packages
         List<PackageInfo> installedPackages = InstrumentationRegistry.getInstrumentation()
                 .getContext().getPackageManager().getInstalledPackages(0);
-        int usedLibraryVersionCode = -1;
+        long usedLibraryVersionCode = -1;
         for (PackageInfo installedPackage : installedPackages) {
             if (STATIC_LIB_PROVIDER_PKG.equals(installedPackage.packageName)) {
                 if (usedLibraryVersionCode != -1) {
